Create CPU resouce service provider
The need for this service provider is felt at this point since we would like to use bounding boxes data in other render passes but they are created in an IndirectRenderer instance. Many more of such use cases will probably be seen in the near future as we try to use the cpu side to store and update some of our resources.