Add utility methods for finding gpu resources
The utility methods that were already present for fetching a gpu resource were insufficient for some cases. Assume we have the case of a gpu resource being created as follows: a texture is created with 2 duplicates: "Texture 1" and "Texture 2" each corresponding to some swapchain with the same index. Then in order to fetch one of the duplicates we would have to first form the corresponding formatted string with the ccorrect index and then fetch the meta data and then call the appropriate retrieve method using the handle in the meta data to get the gpu resource. The added functions do that automatically.