Add half-Z AABB for better perf for 512+ lights
On average the basic tiled deferred we implemented using AABB performs well for point lights under 512. However, due to depth discontinuities we have a lot of false positives with AABBs. In order to make the culling more precise we use half-Z method. For number of point lights under 512 this doesn't show much of a substantial gain in perf; in fact due to extra time spent on culling we get slightly less performance out of this. However, the culling for 512+ lights compared to lightning shows its worth and we get better perf than if we were to resort to normal version.